FIX: function_name change

+++ src/include/parse.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   parse.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jeelee <jeelee@student.42seoul.kr>         +#+  +:+       +#+        */
+/*   By: jhwang2 <jhwang2@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/06 21:14:41 by jeelee            #+#    #+#             */
-/*   Updated: 2023/06/12 16:18:56 by jeelee           ###   ########.fr       */
+/*   Updated: 2023/06/13 19:47:55 by jhwang2          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,11 +43,11 @@ void	setting_object(int type, char *line, t_data *data);
 t_light	**create_llist(int n);
 void	add_llist(t_light *l, t_light ***llist);
 
-int		parse_ratio(char *line, float *ratio);
+int		parse_ratio(char *line, double *ratio);
 int		parse_color(char *line, uint32_t *color);
 int		parse_coordi(char *line, t_point *coodi);
 int		parse_n_vector(char *line, t_point *coodi);
-int		parse_decimal(char *line, float *decimal);
+int		parse_decimal(char *line, double *decimal);
 int		parse_integer(char *line, int *integer);
 
 int		parse_atoi(char *str, size_t size);
